+++ changelog.md
@@ -1,2 +1,5 @@
 ### Fixed
 - `\MvcCore\Tool::GetPascalCaseFromDashed()` in PHP<=5.6,
+
+### Added
+- `\MvcCore\Controller::GetCallerControllerClass()` to get creation class.